Double => int + removed hoeveelheid from manual item, beautified names. Made logging.

+++ Model/interface/model/Recept.h
@@ -17,17 +17,17 @@ public:
   virtual std::tstring          GetName() const                         { return mName; }
   virtual double                GetPoints() const;
 
-  double                        GetNumberOfPortions() const             { return mNumberOfPortions; }
+  int                           GetNumberOfPortions() const             { return mNumberOfPortions; }
   double                        GetPointsPerPortion() const             { return mPointsPerPortion; }
 
-  void                          SetNumberOfPortions(double aNumber)     { mNumberOfPortions = aNumber; }
+  void                          SetNumberOfPortions(int aNumber)        { mNumberOfPortions = aNumber; }
   void                          SetPointsPerPortion(double aPunten)     { mPointsPerPortion = aPunten; }
 
   virtual void                  Accept(ItemVisitor & aVisitor);
 
 private:
   std::tstring                  mName;
-  double                        mNumberOfPortions;
+  int                           mNumberOfPortions;
   double                        mPointsPerPortion;
 };
